refactor: Update AJAX request in create-user.html to send username as JSON
The AJAX request in create-user.html has been updated to send the username as JSON data. This change ensures that the server receives the username in the correct format. Additionally, the content type of the request has been set to 'application/json' to indicate that the data being sent is JSON.