refactor: Add transformToJsDate() function for converting PHP dates to JavaScript dates
The `transformToJsDate()` function has been added to the `utils/transform-to-js-date.php` file. This function converts a PHP date string to a JavaScript date string in ISO 8601 format. This change enables the `create-api.php` file to correctly format the `created_at` and `updated_at` fields when returning user data as JSON.